Made interpreter option that shows a window with initial configuration.

+++ compiler/src/main/kotlin/ast/ast.kt
@@ -302,9 +302,9 @@ class ConstDecl(
 class StateDecl(
     ctx: SourceContext,
     var ident: String,
-    var red: Short,
-    var blue: Short,
-    var green: Short,
+    var red: Int,
+    var blue: Int,
+    var green: Int,
     val body: CodeBlock
 ) : Decl(ctx)
